requirementsMultiScale.txt
certifi==2021.5.30
charset-normalizer==2.0.6
click==8.0.1
cycler==0.10.0
filelock==3.2.0
huggingface-hub==0.0.17
idna==3.2
joblib==1.0.1
kiwisolver==1.3.1
matplotlib==3.3.4
numpy==1.19.5
opencv-python==4.5.3.56
packaging==21.0
Pillow==8.3.2
pyparsing==2.4.7
python-dateutil==2.8.2
PyYAML==5.4.1
regex==2021.9.30
requests==2.26.0
sacremoses==0.0.46
scikit-learn==0.24.2
scipy==1.5.4
six==1.16.0
sklearn==0.0
threadpoolctl==3.0.0
tokenizers==0.10.3
pytorch==1.8.0
torchvision==0.10.0
tabulate==0.8.3
tqdm==4.62.3
transformers==4.11.2
typing-extensions==3.10.0.2
urllib3==1.26.7
